update gitconfig for project management
Create replacement for for-each-repo that prints the repo name before
executing the git command. This means we don't need custom aliases for
every command to print the context.

Create aliases to maintain a single ~/.gitconfig-project file, which
includes a "foreach" alias, a "top directory" definition, and a list of
repos for each project. Optionally, the project can include a specific
~/.gitconfig-* file (to replace the user's identity).